게시글 이미지, 동영상 업로드 기능 진행중

+++ src/main/java/com/shoppingmall/board/repository/CommentRepository.java
@@ -5,14 +5,15 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.shoppingmall.board.model.Board;
 import com.shoppingmall.board.model.Comment;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Integer>{
+public interface CommentRepository extends JpaRepository<Comment, Long>{
 	
 	//전체 최신순 정렬
-	List<Comment> findByBoard_BoardIdOrderByCommentId(int boardId);
+	List<Comment> findByBoard_BoardIdOrderByCommentId(Long boardId);
 	
 	//댓글 수
-	int countByBoardBoardId(int boardId);
+	int countByBoardBoardId(Long boardId);
 }
